datasetresearch_cli/requirements.txt
rich>=13.0.0
click>=8.0.0
colorama>=0.4.4
pyyaml>=6.0
tqdm>=4.64.0
tabulate>=0.9.0